Feat-PlayerEnergy Rework energy system to respond to review
Now the energy pickup is the one which add energy to the player and not the player which add energy when detect energy pickup.
Energy Pickup and Interactible object can be create from the SR_InteractibleObject, when you create the blueprint, you have to add the "Interactable" tag ou "EnergyPickup" tag on creation.